add results='asis' to stargazer chunk header
Added `results='asis'` to stargazer chunk header. As discussed in the lecture slides, this is needed to be able to return the raw code to the compiler so that it can be rendered in the document, rather than as a highlighted output code chunk.

Given that this for your website, I only included the `html` table stargazer call and changed the `output` type to `html_document` in the YAML header. 

Also, RMarkdown uses the `.Rmd` file's current directory as the working directory, therefore you do not need to set the absolute file path for data import calls. Using the relative file path makes it much easier to reproduce your document.